chessgame.js: Make pages more responsive in Chrome
There’s a glitch where Chrome will not render the page at
all until all of the chess diagrams are rendered in Javascript;
this can delay loading of the page by a coupld of seconds and
results in there being a “flash” while Chrome renders the page.

To work around this, the code which renders the chess boards is now
called with window.setTimeout() so that the page can render quickly
sans chess boards, and then render the chess boards after loading
and displaying the page.  As long as the chessboards are “below the
fold”, the only thing the user will notice is that the scrollbar
changes height, but this is normal behavior in an era of infinite
scroll pages (Twitter, Facebook, etc.)

If one does end up where a chessboard is, the board will “pop in”
the existance after being rendered (2-3 seconds at most); while this
is less than perfect, this is better than waiting for all of the
chessboard to be drawn before displaying any of the page.

Since older MSIE doesn’t have window.setTimeout support, we have JS
which checks for older IE; if found, it will just directly render the
chessboards instead of making the chessboard drawing an async event.